Norm: move word expansion
